Reorganize CUDA's future implementation into its own directory
