+++ agency/cuda/detail/future/continuation.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/detail/asynchronous_state.hpp>
+#include <agency/cuda/detail/future/asynchronous_state.hpp>
 #include <agency/detail/integer_sequence.hpp>
 #include <agency/detail/tuple.hpp>
