+++ agency/cuda/detail/new_then_execute.hpp
@@ -2,8 +2,8 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/cuda/detail/on_chip_shared_parameter.hpp>
-#include <agency/cuda/detail/event.hpp>
-#include <agency/cuda/detail/asynchronous_state.hpp>
+#include <agency/cuda/detail/future/event.hpp>
+#include <agency/cuda/detail/future/asynchronous_state.hpp>
 #include <agency/cuda/future.hpp>
 #include <agency/cuda/gpu.hpp>
 #include <agency/detail/factory.hpp>
